requirements.txt
# AI Engine requirements
torch>=1.9.0
transformers>=4.8.0
sentence-transformers>=2.2.0
faiss-cpu>=1.7.0
PyMuPDF>=1.19.0
nltk>=3.6.0
numpy>=1.20.0
scikit-learn>=0.24.0
pandas>=1.3.0
typer>=0.4.0
pdfplumber>=0.7.0
python-docx>=0.8.11
